+++ src/app/search/suggestions-generation.ts
@@ -55,9 +55,23 @@ const operators = ['<', '>', '<=', '>=']; // TODO: add "none"? remove >=, <=?
  * if you want to generate some keywords without a full valid filter
  */
 export function generateSuggestionsForFilter(
+  filterDefinition: Pick<FilterDefinition, 'keywords' | 'suggestions' | 'format' | 'deprecated'>
+) {
+  return generateGroupedSuggestionsForFilter(filterDefinition, false).flatMap(
+    ({ keyword, ops }) => {
+      if (ops) {
+        return [keyword].concat(ops.map((op) => `${keyword}${op}`));
+      } else {
+        return [keyword];
+      }
+    }
+  );
+}
+
+export function generateGroupedSuggestionsForFilter(
   filterDefinition: Pick<FilterDefinition, 'keywords' | 'suggestions' | 'format' | 'deprecated'>,
   forHelp?: boolean
-) {
+): { keyword: string; ops?: string[] }[] {
   if (filterDefinition.deprecated) {
     return [];
   }
@@ -69,35 +83,52 @@ export function generateSuggestionsForFilter(
 
   const allSuggestions = [];
 
+  const expandFlat = (stringGroups: string[][], minDepth = 0) =>
+    expandStringCombinations(stringGroups)
+      .slice(minDepth)
+      .flat()
+      .map((s) => ({ keyword: s }));
+
+  // We delay expanding ops because ops on their own expand the filters list significantly.
+  // For autocompletion `generateSuggestionsForFilter` above expands the ops, but the filters
+  // help has some special display to group the operator variants.
+  const expandOps = (stringGroups: string[][], ops: string[]) => {
+    const combinations = expandStringCombinations([...stringGroups, []]);
+    const partialSuggestions = combinations
+      .slice(0, stringGroups.length - 1)
+      .flat()
+      .map((s) => ({ keyword: s }));
+    const opSuggestions = combinations[stringGroups.length - 1].map((s) => ({ keyword: s, ops }));
+    return partialSuggestions.concat(opSuggestions);
+  };
+
   for (const format of canonicalFilterFormats(filterDefinition.format)) {
     switch (format) {
       case 'simple':
         // Pass minDepth 1 to not generate "is:" and "not:" suggestions. Only generate `is:` for filters help
         allSuggestions.push(
-          ...expandStringCombinations([forHelp ? ['is'] : ['is', 'not'], thisFilterKeywords], 1)
+          ...expandFlat([forHelp ? ['is'] : ['is', 'not'], thisFilterKeywords], 1)
         );
         break;
       case 'query':
         // `query` is exhaustive, so only include keyword: for autocompletion, not filters help
         allSuggestions.push(
-          ...expandStringCombinations([thisFilterKeywords, filterSuggestions], forHelp ? 1 : 0)
+          ...expandFlat([thisFilterKeywords, filterSuggestions], forHelp ? 1 : 0)
         );
         break;
       case 'freeform':
-        allSuggestions.push(...expandStringCombinations([thisFilterKeywords, []]));
+        allSuggestions.push(...expandFlat([thisFilterKeywords, []]));
         break;
       case 'range':
-        allSuggestions.push(...expandStringCombinations([thisFilterKeywords, operators]));
+        allSuggestions.push(...expandOps([thisFilterKeywords], operators));
         break;
       case 'rangeoverload':
-        allSuggestions.push(...expandStringCombinations([thisFilterKeywords, operators]));
-        allSuggestions.push(...expandStringCombinations([thisFilterKeywords, filterSuggestions]));
+        allSuggestions.push(...expandOps([thisFilterKeywords], operators));
+        allSuggestions.push(...expandFlat([thisFilterKeywords, filterSuggestions]));
         break;
       case 'stat':
         // stat lists aren't exhaustive
-        allSuggestions.push(
-          ...expandStringCombinations([thisFilterKeywords, filterSuggestions, operators])
-        );
+        allSuggestions.push(...expandOps([thisFilterKeywords, filterSuggestions], operators));
         break;
       case 'custom':
         break;
@@ -108,15 +139,16 @@ export function generateSuggestionsForFilter(
 }
 
 /**
- * loops through collections of strings (filter segments), generating combinations
+ * loops through collections of strings (filter segments),
+ * generating combinations grouped by number of segments
  *
  * for example, with
  * `[ [a], [b,c], [d,e] ]`
  * as an input, this generates
  *
- * `[ a:, a:b:, a:c:, a:b:d, a:b:e, a:c:d, a:c:e ]`
+ * `[ [a:], [a:b:, a:c:], [a:b:d, a:b:e, a:c:d, a:c:e] ]`
  */
-function expandStringCombinations(stringGroups: string[][], minDepth = 0) {
+function expandStringCombinations(stringGroups: string[][]) {
   const results: string[][] = [];
   for (let i = 0; i < stringGroups.length; i++) {
     const stringGroup = stringGroups[i];
@@ -131,5 +163,5 @@ function expandStringCombinations(stringGroups: string[][], minDepth = 0) {
     );
     results.push(newResults);
   }
-  return results.slice(minDepth).flat();
+  return results;
 }